Bug fixes
- Fix `day` argument when calculating expected break length
- Format break duration as int
- Ignore expected breaks shorter than 5 minutes

+++ app/models.py
@@ -199,7 +199,7 @@ class Break(BaseModel):
 
     @property
     def duration_pretty(self) -> str:
-        minutes = self.duration / 60
+        minutes = int(self.duration / 60)
         return f"{minutes} minutes" if minutes != 1 else "1 minute"
 
     @property
